일부 엔티티 타입 수정(Long -> Integer), useYn, delYn 필드 제거, DTO 수정(세터 제거)

+++ src/main/java/io/hurem/domain/departments/DepartmentsRepository.java
@@ -2,6 +2,6 @@ package io.hurem.domain.departments;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentsRepository  extends JpaRepository<Departments, Long>  {
+public interface DepartmentsRepository  extends JpaRepository<Departments, Integer>  {
 
 }
